web-animations/animation-timeline/animation-timeline-get-animation-players.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation timeline getAnimationPlayers() method tests</title>
<meta name="assert" content="Animation timeline getAnimationPlayers() method returns the set of AnimationPlayer objects associated with this timeline that have associated source content which is current">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#the-animationtimeline-interface">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var players = document.timeline.getAnimationPlayers();
    assert_true(players === null || players.length === 0,
        'AnimationTimeline.getAnimationPlayers() method should return either null ' +
        'or empty set if there are no players');
}, 'Test set of objects returned by AnimationTimeline.getAnimationPlayers() method ' +
    'if there are no players associated');

test(function() {
    var doc = newHTMLDocument();
    var player = newAnimationPlayer(createDiv(this, doc));

    assert_not_equals(doc.timeline.getAnimationPlayers(), null,
        'AnimationTimeline.getAnimationPlayers() method should not return null');
    assert_equals(doc.timeline.getAnimationPlayers(), [player],
        'Set of objects returned by AnimationTimeline.getAnimationPlayers() method ' +
        'should contain one element');
}, 'Test set of objects returned by AnimationTimeline.getAnimationPlayers() method ' +
    'if there is one associated player');

test(function() {
    var doc = newHTMLDocument();
    var player1 = newAnimationPlayer(createDiv(this, doc));
    var player2 = newAnimationPlayer(createDiv(this, doc));

    assert_array_equals(doc.timeline.getAnimationPlayers(), [player1, player2],
        'Set of objects returned by AnimationTimeline.getAnimationPlayers() method ' +
        'should contain two elements');
}, 'Test set of objects returned by AnimationTimeline.getAnimationPlayers() method ' +
    'if there are several associated players');

test(function() {
    var doc = newHTMLDocument();
    var player1 = newAnimationPlayer(createDiv(this, doc));
    var player2 = newAnimationPlayer(createDiv(this, doc));
    var players = doc.timeline.getAnimationPlayers();

    players.forEach(function(player) {
        assert_true(isCurrent(player.source),
            'Source content of the AnimationPlayer returned by ' +
            'AnimationTimeline.getAnimationPlayers() should be current');
    });
}, 'Test that source content of the AnimationPlayer returned by ' +
    'AnimationTimeline.getAnimationPlayers() is current');

async_test(function(t) {
    var doc = newHTMLDocument();
    var animation1 = new Animation(createDiv(this, doc), [], 100);
    var player1 = document.timeline.play(animation1);
    var animation2 = new Animation(createDiv(this, doc), [], 300);
    var player2 = document.timeline.play(animation2);
    var animation3 = new Animation(createDiv(this, doc), [], 200);
    var player3 = document.timeline.play(animation3);

    assert_array_equals(doc.timeline.getAnimationPlayers(), [player1, player2, player3],
        'AnimationTimeline.getAnimationPlayers() should return only players, ' +
        'which source content is current');

    setTimeout(
        t.step_func(function() {
            assert_array_equals(doc.timeline.getAnimationPlayers(), [player2, player3],
                'AnimationTimeline.getAnimationPlayers() should return only players, ' +
                'which source content is current');
        }),
        110
    );

    setTimeout(
        t.step_func(function() {
            assert_array_equals(doc.timeline.getAnimationPlayers(), [player2],
                'AnimationTimeline.getAnimationPlayers() should return only players, ' +
                'which source content is current');
        }),
        210
    );

    setTimeout(
        t.step_func(function() {
            var players = doc.timeline.getAnimationPlayers();
            assert_true(players === null || players.length === 0,
                'AnimationTimeline.getAnimationPlayers() should return only players, ' +
                'which source content is current');
            t.done();
        }),
        310
    );
}, 'Test AnimationTimeline.getAnimationPlayers() returns only players, ' +
    'which source content is current');
</script>
</body>
